bears.misc: Rename to bears.general
Those are generally applicable, we're now sorting everything by
language.
