+++ bears/general/LineCountBear.py
