+++ bears/general/SpaceConsistencyBear.py
