+++ bears/tests/general/KeywordBearTest.py
@@ -5,7 +5,7 @@ sys.path.insert(0, ".")
 import unittest
 from coalib.settings.Setting import Setting
 from bears.tests.LocalBearTestHelper import LocalBearTestHelper
-from bears.misc.KeywordBear import KeywordBear
+from bears.general.KeywordBear import KeywordBear
 from coalib.settings.Section import Section
