+++ bears/tests/general/LineCountBearTest.py
@@ -6,7 +6,7 @@ import unittest
 from coalib.settings.Section import Section
 from coalib.results.Result import Result, RESULT_SEVERITY
 from bears.tests.LocalBearTestHelper import LocalBearTestHelper
-from bears.misc.LineCountBear import LineCountBear
+from bears.general.LineCountBear import LineCountBear
 
 
 class LineCountBearTest(LocalBearTestHelper):
